switched type of mem_map.size to unsigned int

+++ varedit.c
@@ -30,7 +30,7 @@ bool mem_rgn_warn(int d_rgn, struct mem_rgn mem, bool additional){
 
 int remove_volatile_values(struct mem_map* vmem){
       int n = 0;
-      for(unsigned long i = 0; i < vmem->size; ++i){
+      for(unsigned int i = 0; i < vmem->size; ++i){
             for(int in = 0; in < 10; ++in){
                   if(vmem->mmap[i].value != read_single_val_from_pid_mem(vmem->pid, 4, vmem->mmap[i].addr)){
                         vmem->mmap[i--] = vmem->mmap[--vmem->size];
@@ -43,7 +43,7 @@ int remove_volatile_values(struct mem_map* vmem){
 
 void print_mmap(const struct mem_map* mem, const char* contains, bool integers, bool show_rgns){
       char tmp_num[20];
-      for(unsigned long i = 0; i < mem->size; ++i){
+      for(unsigned int i = 0; i < mem->size; ++i){
             if(integers){
                   sprintf(tmp_num, "%d", mem->mmap[i].value);
                   if(strcmp(contains, "") == 0 || strcmp(tmp_num, contains) == 0){
@@ -61,13 +61,13 @@ void print_mmap(const struct mem_map* mem, const char* contains, bool integers, 
 }
 
 void logic_swap(const struct mem_map* mem){
-      for(unsigned long i = 0; i < mem->size; ++i){
+      for(unsigned int i = 0; i < mem->size; ++i){
             if(mem->mmap[i].value == 0 || mem->mmap[i].value == 1)
             write_bytes_to_pid_mem(mem->pid, 1, mem->mmap[i].addr, (unsigned char*)&mem->mmap[i].value);
       }
 }
 
-bool interactive_mode(struct mem_map* vmem, bool integers, int int_mode_bytes, int d_rgn, int additional, bool verbose, unsigned long result_print_limit, bool print_rgns){
+bool interactive_mode(struct mem_map* vmem, bool integers, int int_mode_bytes, int d_rgn, int additional, bool verbose, unsigned int result_print_limit, bool print_rgns){
       char search_mode_help[600];
       strcpy(search_mode_help, "search mode options:\n    'r' : reset mem map\n    \"wa\" <value> : write single value to all current results\n    ");
       if(integers)strcpy(search_mode_help+110, "<integer> : enter an integer to narrow results\n    \"rv\" : remove volatile variables\n    ");
@@ -147,11 +147,13 @@ bool interactive_mode(struct mem_map* vmem, bool integers, int int_mode_bytes, i
                               tmp_val = atoi(tmp_str+3);
                               memcpy(write, &tmp_val, int_mode_bytes);
                         }
-                        for(unsigned long i = 0; i < vmem->size; ++i){
+
+                        for(unsigned int i = 0; i < vmem->size; ++i){
                               if(integers)write_bytes_to_pid_mem(vmem->pid, int_mode_bytes, vmem->mmap[i].addr, write);
+                              // TODO allow writing \0 in wa mode
                               else write_str_to_pid_mem(vmem->pid, vmem->cp_mmap[i].addr, tmp_str+3);
                         }
-                        printf("wrote \"%s\" to %li memory locations\n", tmp_str+3, vmem->size);
+                        printf("wrote \"%s\" to %i memory locations\n", tmp_str+3, vmem->size);
                         goto Find;
                   }
             }
@@ -168,16 +170,16 @@ bool interactive_mode(struct mem_map* vmem, bool integers, int int_mode_bytes, i
                               goto Find;
                         }
                         if(integers){
-                              for(unsigned long i = 0; i < vmem->size; ++i){
-                                    printf("%li: (%p: %i)\n", i, vmem->mmap[i].addr, vmem->mmap[i].value);
+                              for(unsigned int i = 0; i < vmem->size; ++i){
+                                    printf("%i: (%p: %i)\n", i, vmem->mmap[i].addr, vmem->mmap[i].value);
                               }
                         }
                         else{
-                              for(unsigned long i = 0; i < vmem->size; ++i){
-                                    printf("%li: (%p: \"%s\")\n", i, vmem->cp_mmap[i].addr, vmem->cp_mmap[i].value);
+                              for(unsigned int i = 0; i < vmem->size; ++i){
+                                    printf("%i: (%p: \"%s\")\n", i, vmem->cp_mmap[i].addr, vmem->cp_mmap[i].value);
                               }
                         }
-                        printf("enter a number from [0-%li] or a range with a '-', followed by a value to write OR 's' to continue searching\n", vmem->size-1);
+                        printf("enter a number from [0-%i] or a range with a '-', followed by a value to write OR 's' to continue searching\n", vmem->size-1);
                         // width is 1 less than length of length of v_loc_s to avoid overwriting '\0'
                         // ignore leading whitespace
                         scanf(" %9[^ \t.\n]%*c", v_loc_s);
@@ -366,7 +368,7 @@ bool interactive_mode(struct mem_map* vmem, bool integers, int int_mode_bytes, i
             }
             else{
                   if(!verbose && vmem->size > result_print_limit){
-                        printf("your search of %s has %li results\nresult_print_limit is set at %li. refusing to print\n", tmp_str, vmem->size, result_print_limit);
+                        printf("your search of %s has %i results\nresult_print_limit is set at %i. refusing to print\n", tmp_str, vmem->size, result_print_limit);
                   }
                   else{
                         puts("matches are now:");
